+++ src/components/entities/entity.js
@@ -5,8 +5,6 @@ import random from 'lodash.random'
 
 import materials from 'world/materials'
 
-import mixin from 'utils/mixin'
-
 /**
  * Entity is a fairly abstract class, usually it’ll be instantiated with either
  * a body (ShellEntity), a renderable (GhostEntity) or both (PhysicalEntity).